[ot-stack] Updating files to support soft-transition
A minor change to radio Init function is needed to accept 3 arguments
for new version of files from openthread.git whereas 2 args are needed
in current version of openthread.git.

Added third argument conditionally based on a gn variable which is
defind in a gni file under third_party/openthread. Using this method
the change compiles with both old and new openthread.git changes.

Functionally it is a no-op as the value 'false' of variable is
equivalent in functionality to 2 variable case.

Bug: 73064
Tests:
       fx test threadstack-unittests
       fx test threadstack-bootstrap-unittests

       functionally it's a no-op as it is exact same compilation with
       current ot-stack.

Change-Id: I536b7ff51f6bdcfef3a18c23b2b7e67c4a7f7ed2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/506882
Reviewed-by: Robert Quattlebaum <rquattle@google.com>
Commit-Queue: Parag Dixit <paragdixit@google.com>