[ffx] ensure version information is generated
Previously the script did not fail when the version information could
not be gathered, and a recent change to the integration repository
resulted in HEAD being ambiguous to git show in some user scenarios.
We now rev-parse and then show that commit to generate our string,
whcih is more in-line wiht the behavior that the Zircon git hash
collection uses.

The dependencies are changed to only depend on HEAD as well, as a
small speculative change to try to somewhat reduce the frequency of
needless incremental rebuilds in infrastructure. This again shares
behavior with the Zircon git hash computation. A more complete
solution should follow as described in the associated bug.

Bug: 73152
Change-Id: Ic5ead447ea4b9d1ce69de24d0a8f0feb053de93e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/507325
Fuchsia-Auto-Submit: James Tucker <raggi@google.com>
Reviewed-by: Jordon Wing <jwing@google.com>
Commit-Queue: James Tucker <raggi@google.com>