[component_manager] CapabilityProvider doesn't Task::detach()
CapabilityProvider is a trait that component manager modules can use to
implement builtin or framework capabilities. It is common for trait
implementations to spawn tasks with detach().

Instead of calling detach(), have CapabilityProvider return the task and
stash it in an appropriate scope. For scoped framework capabilities,
store it in the instance of that scope. This will cause the task to get
cancelled if the scope instance is destroyed. For builtin capabilities,
store the task in component manager's instance.

This is one step toward removing use of Task::detach() in component_manager.

Change-Id: Icaaae6d7d1a8a759d3e58418e362a615f7c0850f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/501960
Commit-Queue: Gary Bressler <geb@google.com>
Reviewed-by: Xyan Bhatnagar <xbhatnag@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>