[block] Fix use-after-free in MessageGroup
MessageGroup was unsafely using `delete this` while holding a lock on
this, which resulted in a UAF when the lock guard went out of scope.

Fixed this by refactoring so we don't need to use `delete this` to clean
up the MessageGroup.

Bug: 73044 # Fixed
Change-Id: I88a8c44f4caa839a91e62800c991bcb6068162d7
Tested: core-unittest.cmx with ASAN enabled.
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/507199
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Brett Wilson <brettw@google.com>