[fx-set] Generate compile_commands.json + rust-project.json
The legacy version of `fx set` passes
`--export-compile-commands=default` and `--export-rust-project` to GN,
which I forgot to add to the new version of fx (I was mainly concerned
about checking that the new version produced the same GN args, and
neglected to check that it produced the same GN flags).

fint already has support for generating compile_commands.json via the
`generate_compdb` static.proto field, but we need to be able to specify
the "default" target to avoid the latency of generating compile commands
for the entire build graph, so I added a new `compdb_targets` field to
override the set of targets that were include in the compdb.

Bug: 73231
Change-Id: I1416642677b2c6caa889de116e3d5462709de3dc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/508038
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Gary Boone <gboone@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>