[kernel][vm] Clear split bits of forked markers
When performing a fast merge of a parent and remaining child page lists
we only need to update the pages that the removed VMO forked.
Previously markers in removed VMO would be skipped, when in reality a
marker existing indicates a page fork happened, same as if a page is
present. Failing to update the split bits in this case would result in
a consistency error, manifesting only if another clone is created in
the future, as demonstrated by the included test case.

Bug: 72141
Change-Id: Iafafe64b657a70d8df4fa49edfad5d3f3da69267
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/507900
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>