[kernel][arm64][feature] save the ASID bit size
Instead of panicing when booting on an 8 bit asid cpu, save the feature
and print a warning instead.

It's still not supported to really run for any length of time on an 8
bit ASID cpu but it should at least make an attempt at booting.

Save the feature bits for future code to switch behavior based on it.

Change-Id: Icb6c517d070ab32ee71d7c0be98cb6c41340c637
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/507335
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Travis Geiselbrecht <travisg@google.com>