[blobfs] Remove thread from paging test.
The paging test ran many blobfs commands on a background thread, while
simultaneously accessing Blobfs and Blobfs on the main thread. Blobfs is
not designed for this. Fortunately, the background thread seems to be
unneeded and we can run the loop on the main thread.

Bug: 72754
Change-Id: Ifa9974084ac7a64cdd179659a83f716d4d900d2d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/508065
Reviewed-by: Martin Lindsay <mlindsay@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>