[fx set] Rebuild fx-set after HEAD changes
If we only rebuild when JIRI_HEAD changes, `fx set` won't be rebuilt
after checking out an older version of fuchsia.git using raw `git
checkout`. This may result in a failure if the built version of `fx set`
is incompatible with the checked-out fuchsia.git revision.

To handle that case, rebuild whenever HEAD changes, rather than
only when JIRI_HEAD changes.

Note that using raw git to check out and build old versions of
fuchsia.git is not an officially supported workflow, as old versions of
fuchsia.git may not be compatible with up-to-date versions of other
repos and CIPD packages pinned in global integration. However, that
workflow is still useful and generally "just works" as long as you don't
go back in history more than a few days. So we might as well make it
less likely to break, as long as it's easy. It would also be confusing
for a tool built from source in fuchsia.git to not auto-update when
switching to a different version of its sources.

Two potential future improvements:
1. Skip rebuilding if HEAD is *ahead* of "origin/master" (or
   "origin/main" in the future), unless the --dev flag is set. This will
   avoid rebuilding in the case where the user has made local changes
   unrelated to "fx set".
2. Maintain a cache of old versions of `fx-set` keyed by revision to
   avoid rebuilding when switching back and forth between two branches
   multiple times. But then we'd have to deal with garbage collection
   and such, which I'd rather not think about unless it's absolutely
   necessary.

Bug: 68465
Change-Id: I1c8827dc48755d996642a228b250c2c11c2cea01
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/508219
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Gary Boone <gboone@google.com>