[kernel][arm64][mmu] simple implementation of 8 bit ASID support
Simple implementation: tweak the asid allocator to just wrap at 256 and
keep selecting ASIDs in the 8 bit range instead of 16 bit. It's not a
perfect solution since it effectively limits the system to 254 processes
before running out, but its better than looping around and blowing up.

Had to add some more plumbing to let the asid allocator object get
constructed after the feature bits are parsed. Added a debug assert in
the asid width reader code to make sure some restructuring of the code
in the future doesn't cause it to get constructed prior to the feature
bits.

Added simple unit tests to make sure the allocator runs out of bits when
it should.

Bug: 71226

Change-Id: I57f7947e2a5f696f4289c31772197e108a2790dd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/507404
Commit-Queue: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>