[blobfs] Fix some problems with new pager.
Blobfs::Create will actually make the Blobs. But before this patch, the
Vfs was not set on the Blobfs object until after Create was complete.
This meant that none of the created blobs got a Vfs pointer, resulting
in the pager failing.

Creation is changed to take the Vfs as an input pointer. A set function
is added on the Runner object to link the two sides (this was done in
the reverse order before).

Additionally, PagedVfs creation is improved to avoid reentrancy created
by the paged blobs being deleted after the destructor was complete.

Bug: 51111

Change-Id: I29649855db8081126aee97b552b459b8b6f7f879
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/506563
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Martin Lindsay <mlindsay@google.com>