[kernel][stacks] Allocate a vmo only large enough to hold the stack
The inner kstack allocation and mapping routine was previously always
allocating DEFAULT_STACK_SIZE vmos instead of the size passed in on the
particular stack type descriptor. In the case of shadow call stacks on
arm this means the vmo was actually sized at 8192 bytes instead of 4096
as was needed.

This was basically harmless because only the mapped portion of the vmo
is faulted in as the mapping is created. However some future code may
want to fiddle with additional types and would immediately notice if the
size was greater than DEFAULT_STACK_SIZE.

Cleaned up the kstack implementation to encapsulate the inner stacks
into a repeated structure.

Change-Id: Id7349e0a56fd54d690774e8a6259071121c26853
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/418117
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Travis Geiselbrecht <travisg@google.com>