[weave] Add resource to WeaveTestFixture
To ensure that resource lifetimes outlast the test fixture loop, this
change adds a resource to the WeaveTestFixture class, which will ensure
that the resource will remain alive until TearDown is completed, and
will also ensure that the resource will be deleted after TearDown is
completed.

Additionally, this change applies a resource to the existing
ConfigurationManager tests to ensure that created PseudoDirs outlive the
fixture loop.

MULTIPLY: `[
  {
    "name": "configuration_manager_unittests",
    "total_runs": 1000
  }
]`

Bug: 72082
Test: fx test weavestack-adaptation-unittests
Change-Id: I1ed2bd1f30abab91523dd742b260a5922d65007f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/502038
Commit-Queue: Jason Graffius <jgraff@google.com>
Reviewed-by: Prashanth Swaminathan <prashanthsw@google.com>
Reviewed-by: Neethu Joseph <neethujoseph@google.com>
Reviewed-by: Sundarajan Srinivasan <sundarsrini@google.com>