[sdk][stdcompat] Temporarily disable test
This test is blocking the clang roll from an update to __cpp_lib_variant
which reflects changes to std::variant. We will temporarily disable it
to facilitate the clang roll, but should re-enable it after the roll and
assert the new std::variant works with polyfill.

Bug: 73340
Change-Id: I024c71e180f7dfbe4b9b9cbc61bfbabe14114946
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/508705
Reviewed-by: Petr Hosek <phosek@google.com>
Commit-Queue: Leonard Chan <leonardchan@google.com>