[testrunner] Collect data sinks from component v2 tests.
This adds a RemoteDir field to the DataSinkReference to distinguish
between the two different directories where the data will be written to
for v1 versus v2 tests. It also collects all v2 data sinks at the end of
all tests because sometimes it takes some time for the summary.json to
get written to so we may fail to get the data sinks for the first test
if we try to read the summary.json before it's written.

Change-Id: Ia4e976a40dc7f947a138f12c65ea991e87d77567
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/503538
Commit-Queue: Ina Huh <ihuh@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>