[bt][hfp] Identify calls by index instead of phone number.
In the HFP procedures, Calls are referred to by an assigned call index
instead of by the phone number of the remote party.

Introduces a new data structure, `CallList` which has special operations
for insertion, retrieval, and deletion so that the index associated with
a given call conforms to the standard.

Bug: 64563
Test: Add unit tests

Change-Id: Iea95caf35a97bf3d56b9433b9c929b5bbc908048
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/497388
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Jeff Belgum <belgum@google.com>
Reviewed-by: Matthew Kehrt <kehrt@google.com>