[kernel][oom] Remove redundant eviction trigger
Do not schedule asynchronous eviction for the future when the pressure
level is OOM. We perform synchronous eviction at OOM level instead.

Also, filesystems are shut down at OOM, so any eviction beyond this
point will be a no-op, as there won't be any pager-backed memory to
evict. Once we've performed synchronous eviction, and still decide to
send the OOM signal and reboot because free memory is below the OOM
threshold, the OOM state is final.

Test: Manual testing with `k pmm mem_avail_state` and `k pmm oom`.

Change-Id: I3e98257ea96abd4237efc55eddf2bc149a08e229
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/508348
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>