[kernel][pmm] Atomic for free count
The free_count_ is currently read racily without a lock. Although this
is unlikely to cause a correctness issue, it is still undefined
behavior. In this change it is made an atomic with all atomic accesses
using the relaxed memory ordering, with the locking itself left alone.

Change-Id: Ia4e01d50eefa7e1d043185c4fc01043f90f93517
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/508438
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>