[fidlgen][cpp] #ifdef __Fuchsia__ handling
It turns out that the namespace stack was only useful and used for
managing when entering and leaving sections of generated code that may
only be compiled on Fuchsia. This has now been reworked so that the
PushNamespace and PopNamespace template functions become IfdefFuchsia
and EndifFuchsia template functions and handle generating ifdef and
endif preprocessor directives (if needed).

Change-Id: Ib512f2bc53b667d97b8dedbfec0dcee8913b3180
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/508238
Commit-Queue: Ian McKellar <ianloic@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>