[bt][hfp] Add format type method to Number.
The format of the number should be associated with the Number itself.
Number is moved to a submodule so that Numbers must be constructed
with methods which can verify the format and modify the string value.

All Numbers must include enclosing double-quotes which are added.

NUMBER_FORMAT is chosen to be within the range 128-143 which indicates
an international or national number that should be presented as is.

Test: Added unit tests.

Change-Id: I1723700bc7c7f2304bb7489093b34a90a4545efe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/508549
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Jeff Belgum <belgum@google.com>
Reviewed-by: Matthew Kehrt <kehrt@google.com>