[fxfs] Add FxfsServer
FxfsServer is the runtime for an fxfs instance, which listens to the
various FIDL interfaces and routes requests to the appropriate VFS node.

This work also revealed some issues with shutdown. Since the
BlockDevice::close implementation needs to unregister its VmoId,
BlockDevice::close either needed to take &mut self (which infects the
next level up with mutability, and gets quite complicated given that we
mainly use Arc<dyn Device> which can only be turned into a &mut dyn
Device if there's only one reference), or VmoId needs to be interiorly
mutable to permit BlockDevice to take its inner value in close(). I've
opted for the latter and changed VmoId to wrap an AtomicU16.

Tested: Lifecycle unit test.

Change-Id: I92cd9fb3dc372ebdd41c9f77c91b7c6de246c503