[kernel][phys] Ensure %gs is set up to point to beginning of TLS area.
In phys _start, we currently point %gs to the end of the TLS section,
instead of the beginning as required.

In particular, x86-64 uses the offsets:

  #define ZX_TLS_STACK_GUARD_OFFSET 0x10
  #define ZX_TLS_UNSAFE_SP_OFFSET 0x18

for TLS, both of which require %gs to point to the beginning of the
TLS area, not the end.

Change-Id: I1b923997f70be6189d2232d494068238a3455074
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/508578
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Fuchsia-Auto-Submit: David Greenaway <dgreenaway@google.com>