[update package] Correct the tracking of sources used in the update package
Correctly specify the set of files that are used to construct the package as sources
of the action that creates the  package.

Previously, a variable called 'sources' was being used as a temporary to unpack the
sources list, for each target to add to the update package.  This was overwriting the
action's own sources list (set above, on line 2174.

This change switches the specification of an update package's entries to use strings
instead of lists of strings, removing the need for the temporary with sources (but
still needs to do the same for any update package entries that are specified by
dependency).

Further, if the build version is unspecified, and using the latest-commit-date file,
the generation of that file is added as a dependency.

Change-Id: Iac90bca56e5428163db8fb7f425f464ad85aebea
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/508210
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Shai Barack <shayba@google.com>