[fidl][rust] Report epitaphs on one-way message call failure
This CL makes one-way method calls report epitaphs if writing to the
channel fails with PEER_CLOSED. This works for an epitaph that was
already discovered and cached (e.g. by a previous method call), and for
an epitaph still in the channel waiting to be read.

This does not affect performance in the success case. In the PEER_CLOSED
case, one-way method calls will potentially do all the work in recv_all,
(acquire the epitaph lock, read from the channel, decode the epitaph,
wake pending futures). For comparison, two-way method calls already do
this -- in the synchronous portion, before returning a future.

As part of this change, all PEER_CLOSED errors on the client side are
reported via Error::ClientChannelClosed, not ClientWrite or ClientRead,
fixing fxbug.dev/73389.

This also adds tests to ensure that an epitaph discovered during one
operation (e.g. getting the next event) is reported again during another
operation (e.g. sending a one-way method), for all combinations.

Test: fx test fidl_rust_lib_tests
Test: Manually checked that new tests fail as expected before the change
Fixed: 72935
Fixed: 73389
Change-Id: Id6db69d940cfca1ffdd463b23e24ca861eff4a5b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/506573
Reviewed-by: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Marie Janssen 💖 <jamuraa@google.com>
Reviewed-by: Kevin Wells <kevinwells@google.com>
Reviewed-by: Adam Lesinski <adamlesinski@google.com>
Commit-Queue: Mitchell Kember <mkember@google.com>