[zbitl] Have iterator hold the current item's offset
Currently it holds that of the next item. This means that the
determination of item and payload offsets is done by working backwards,
which assumes that headers and payloads are alternatingly laid out and
that headers are of a fixed length; support for the bootfs case will
invalidate these assumptions.

Bug: 68585
Change-Id: I86b67b4402f449748940fa869344db06c86b08f1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/508239
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: David Greenaway <dgreenaway@google.com>