[pkg-cache] Allow unexpected blob closes
Relax the NeededBlobs protocol requirements a bit to allow callers to
close a blob while writing it and later re-open the blob for write, and
fill in some missing test coverage for this case.

The package resolver handles blob fetch errors by closing the blob and
starting over from scratch, and the pkg-cache should continue to allow
that behavior when the pkg-resolver switches to using pkg-cache.

Bug: 64627
Change-Id: If1a244c7de5ee07f62761204d990a4b9857caaaa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/508356
Reviewed-by: Devon H. O'Dell <dhobsd@google.com>
Reviewed-by: Dan Johnson <computerdruid@google.com>
Commit-Queue: Kevin Wells <kevinwells@google.com>