[pci] Enhanced legacy interrupt support
This change contains the majority of the implementation for RFC-0070.

When a userspace PCI_IRQ_MODE_LEGACY IRQ is triggered the bus driver's
IRQ worker will now disable that driver's IRQ generation ability until
the interrupt is handled and acked. This is functionally similar to how
the kernel PCI driver handles this, except it must be done in userspace.

A new legacy IRQ mode, PCI_IRQ_MODE_LEGACY_NOACK, has been added. In
this mode the bus driver will allow up to a configured number of
interrupts per second from a given device before determining a device
has too many spurious IRQs. In the event that this happens the device's
IRQ generation will be disabled until the device changes IRQ modes. In
this CL that limit is 512 interrupts a second, or an interrupt every 2
milliseconds. This presently is set at bus level, but can be both
adjusted and moved to per-device constraints in the future if testing
finds it necessary. Since high performance devices will be using MSI
or MSI-X typically, we may find this limit to be acceptable in practice
provided the bus driver's CPU usage stays low.

Bug: 32978, 71188
Change-Id: I24e9d15b54d85b692b11d1a76d69e4a7b16cd630
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/472848
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Sam Hansen <hansens@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Christopher Anderson <cja@google.com>