[io2] Make fuchsia.io.AddInotifyFilter match fuchsia.io2.AddInotifyFilter
This changes the path from a string to path-specific alias defined in
io2.

The string limit is one shorter, noted as desirable at
https://cs.opensource.google/fuchsia/fuchsia/+/master:sdk/fidl/fuchsia.io/io.fidl;l=288
so this is strictly wire-incompatible, but seems unlikely to cause a
problem in practice.

Source-wise, the bindings treat the path as either a string view, or a
string, not a fixed sized buffer, so also shouldn't be a problem.

Additionally, copy reformatted comment from io2.

Bug: 71365
Change-Id: I33fb23e4c235bb472d3cc0ce9076db4f473f5910
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/504531
Reviewed-by: Adam Barth <abarth@google.com>
Reviewed-by: Manali Bhutiyani <manalib@google.com>
API-Review: Adam Barth <abarth@google.com>
API-Review: Manali Bhutiyani <manalib@google.com>
Commit-Queue: Scott Graham <scottmg@google.com>