[wlan] Remove dependency between WlanHasher, SsidFmt, and MacFmt
There is no longer a dependency between WlanHasher, SsidFmt, and
MacFmt. In addition to the simplification of each module and making it
even easier to remove WlanHasher when the time comes, this also
removes a potential runtime failure path from MacFmt and SsidFmt. The
use of `truncate()` in both traits would have panicked if an
unexpected string encoding was used.

Test: fx test wlan-common-tests wlan-sme-tests
Fixed: 73135
Change-Id: Ied8fda1c39bbee5dd79d138b9a4ded8d5b82718f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/507153
Commit-Queue: Charles Celerier <chcl@google.com>
Reviewed-by: Sean Olson <seanolson@google.com>