[bt][a2dp] Fix cobalt metrics reporting
Metrics reporting often is failing because we are not detaching the
task, and when the Media task is stopped by the stream it is often just
dropped.

Detach the cobalt reporting for now to make sure it submits it's report
to the Cobalt server.

Bug: 71825
Test: added dropped_task_reports_metrics

Change-Id: I400bc44d4ec0286efd3f1f01ff40cdab6c91b620
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/508879
Fuchsia-Auto-Submit: Marie Janssen 💖 <jamuraa@google.com>
Reviewed-by: Zac Bowling <zbowling@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>