[brcmfmac] Avoid debug asserts on rxglom failure
If RX glomming failed to parse the received header, because of a
transfer error or something else, the code would attempt to use
uninitialized values from that parse. This was harmless in release
builds because the computed values were never used but in debug builds
they trigger asserts that crash the driver.

Check the return code of the header parsing and only do the computations
if the parsing succeeded.

Bug: 72260
Test: Not able to reproduce CRC error but verified normal operations
Change-Id: Ic431db5b6c7b4ff68c81be9ef1067062f4627296
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/508838
Reviewed-by: Joshua Conner <joshconner@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>