[fshost] shut down cleanly and mark process as critical
This cl marks fshost as a critical process to the root job through
component manager and fixes some issues that were stopping it from
exiting cleanly during shutdown.

The issues with fshost shutdown -

1. The signal processing code for BlockWatcher overlooked the ShutDown
signal, and as such the destructor blocked forever waiting for the
watcher to exit after the signal was sent.

2. In FsManager::Shutdown, instead of calling UninstallAll, we should be
calling Shutdown so that all the connections to the vfs structure we
host are actually closed before we destroy the vfs objects.
Additionally, we need to shut down both the outgoing_vfs and the
root_vfs to terminate all connections to both fshost services and the
hosted filesystems respectively.

3. Even if the delayed outdir hasn't started executing yet, it is
accumulating connections that will need to be destroyed. This changes
the destructor to always tear down these connections, even if it was
never started.

4. The cobalt metrics thread was trying to flush to cobalt on shutdown,
but the service probably doesn't exist, so it blocked forever while we
tried to join the thread in the destructor. Instead, detach the thread
and remove the shutdown flushing logic.

Additionally, a new test-fshost.cml file is added because we can't set
main_process_critical in test environments. fxbug.dev/73189 tracks
cleaning that up so the two cml files don't fall out of sync.

Bug: 47682
Tests: fx test fshost-tests and fx test dm_reboot_test
Change-Id: Ie7e8a882f101eabe867738e9dc72447f0e102851
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/473058
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Reviewed-by: Bryan Henry <bryanhenry@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Stephen Demos <sdemos@google.com>