[volume_image]: Introduce command adapters.
Each of these commands will map to an existing legacyy tool command and
options. Will make testing the behavior easier. Later will introduce
argument to parameter conversion, which can also be tested,
which will make the binary a trivial call to parse options and command
and a call into the command adapter.

Tests are intentional superficial, in the sense that the library logic
is tested, but just checking that the options are passed.

To use fvm-check, wrapped in if def a section that tries to map to some
fuchsia only code, to figure out the name of the partition. This allows
the tests to run in host and fuchsia.

Test: storage-volume-image-test(host),
storage-volume-image-tests(fuchsia),
storage-volume-image-adapter-integration-tests

Change-Id: I24c883e004687d0707918026978a5eebc3f762b6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/507783
Reviewed-by: Chris Suter <csuter@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>