[starnix] Generate uapi using bindgen
This CL uses bindgen to automatically generate some of the constants and
types from uapi that we had previously typed in manually.

Currently, we need to run the bindgen tool manually using a script.
Hopefully we'll be able to run bindgen as part of the build in the
future.

We seem to be using a different signed/unsigned type for various
constants than what bindgen generates. In this CL, I've left put in a
manual conversion in types.rs, but we should study whether we can use
the signed-ness that bindgen generates without needing the conversion in
a future CL.

Bug: 70902
Change-Id: Ifbd78558b0b15d5defefeb23fbac80181de9bc8f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/509810
Commit-Queue: Adam Barth <abarth@google.com>
Reviewed-by: Theodore Dubois <tbodt@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>