[fidlgen][cpp] Migrate fuzzer generators to use CodegenOptions
The fuzzer bindings have so far hardcoded the include stems of the C++
bindings which it depends on. This CL changes it to instead take the
include stems from the command line.

A nice effect is that all include stem source-of-truths are now defined
in fidl_cpp.gni, making it easier for us to change them later.

We are also able to remove the FuzzerHeaders field in the IR, which used
to simply duplicate Headers. This was because the fuzzer codegen was
overwriting Headers with its own values to point to the HLCPP/LLCPP
bindings. After making the binding include stem first-class, we are able
to simply use Headers to represent fuzzer headers, same as all other
binding flavors.

Change-Id: Ia0c2c730c6ed1873cdfb931f4f20b0c041c6bfd3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/508542
Commit-Queue: Yifei Teng <yifeit@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Reviewed-by: Mark Dittmer <markdittmer@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>