[Bind] Fix symbol encoding in new bytecode
Modify the mismatch symbol type check in the new
bytecode encoding. Since the LHS symbol represents
the property key and the RHS symbol represents the
property value, the symbol types might not match.

Instead, the code now checks if the LHS symbol
represents a key and the RHS symbol represents a
value. If the LHS symbol contains a value type,
it is compared with the RHS symbol to ensure that
they match.

Bug: 71051
Test: Updated and added unit tests
Change-Id: I3525c3c369718a0785bd83aa6b6023802b758abc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/505938
Reviewed-by: Alex Legg <alexlegg@google.com>
Commit-Queue: Sarah Chan <spqchan@google.com>