[session] Don't connect to Launcher in element management tests
SimpleElementManager::new() connects to to the Launcher protocol
in the component's incoming directory. This protocol is not routed
in tests (and it shouldn't be), resulting in logspam, e.g.

    [00496.513644][108203][108380][<root>] WARN: No capability
    available at path /svc/fuchsia.sys.Launcher for component
    /tests:auto-9262243642210911116:24/test_wrapper:0/test_root:0,
    verify the component has the proper `use` declaration.

This also causes tests that use `run_until_stalled` to flake if
the test body finishes before the connection fails:

  [00496.513405][1195728][1195730][<root>][element_manager_tests.cm]
    INFO: launch_v2_element(...)
  [00496.513561] ... ERROR: PANIC: Stalled without completing. ...
  [00496.513644] ... WARN: No capability available at path ...

Fixed: 73241
Change-Id: I55ee9bbd0640ae5f16285ac43bd2f36b5a7e5d32
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/509143
Commit-Queue: Yegor Pomortsev <ypomortsev@google.com>
Reviewed-by: Rich Kadel <richkadel@google.com>