[bt][snoop] Use monotonic clock when UTC is not available.
Store all packet timestamps in clock monotonic time. Apply
a transformation to UTC when serializing packets to FIDL or
Inspect. Keep timestamps in clock monotonic, if UTC is not
available when serializing.

FIDL clients that are already following the packet logs when
the UTC clock becomes available will continue using clock
monotonic for the remainder of the session. This is done to
avoid a discontinuity in timestamps.

Bug: 72835
Test: unittests for clock.
Test: manually checked wireshark output is correct.
Change-Id: I32c32f446e1323ef270a3b0b86ea8429c18d4c67
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/507782
Fuchsia-Auto-Submit: Jeff Belgum <belgum@google.com>
Reviewed-by: Marie Janssen 💖 <jamuraa@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>