[sdk][cts] Add "cts" Category of SDKs
This change adds a "cts" category of SDKs. This enables the creation and
release of a CTS-specific SDK, which can be used as a mechanism to
release CTS sources. A CTS-specific SDK would be used to release
libraries required to build the CTS, but should not be released /
depended on in a public/partner SDK.

Bug: http://fxbug.dev/72547
Change-Id: I24240c13b7d8967cc073cdc171c46ab1491eb797
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/506838
Reviewed-by: Misha Gridnev <gridman@google.com>
Commit-Queue: John Shamoon <johnshamoon@google.com>