[wlan][sl4f] always set client controller
Previously, if the facade already had a client controller when
create_client_controller was called, the facade would not replace the
client controller. However, if something has the client controller when
create_client_controller is called, we may not get an error until we try
to use the facade's client controller. The previous implementation would
not try to create a new client controller if the facade already has one,
so there would be no way to get a working controller in this situation.

This implementation will always drop the current controller and then
create a new controller.

Bug:71244
Test: Ran a policy ACTS test and verified that it still works:
	act.py -c <path to config> -tc SavedNetworksTest
      Manually tested that calling create_client_controller twice in an
      ACTS test would not cause an error.

Change-Id: I7412293416523459bb88796d2d360e394f26b48d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/509473
Reviewed-by: Hayden Nix <haydennix@google.com>
Commit-Queue: Naomi Mccracken <nmccracken@google.com>